- Add TestHelper and move redundant code there
